+++ packages/runtime/src/presets/preset-attributify.ts
@@ -1,8 +1,6 @@
 import type { AttributifyOptions } from '@unocss/preset-attributify'
 import presetAttributify from '@unocss/preset-attributify'
 
-(() => {
-  window.__unocss_presets = Object.assign(window.__unocss_presets ?? {}, {
-    presetAttributify: (options: AttributifyOptions) => presetAttributify(options),
-  })
-})()
+window.__unocss_presets = Object.assign(window.__unocss_presets ?? {}, {
+  presetAttributify: (options: AttributifyOptions) => presetAttributify(options),
+})
