+++ packages/runtime/src/presets/preset-mini.ts
@@ -1,8 +1,6 @@
 import type { PresetMiniOptions } from '@unocss/preset-mini'
 import presetMini from '@unocss/preset-mini'
 
-(() => {
-  window.__unocss_presets = Object.assign(window.__unocss_presets ?? {}, {
-    presetMini: (options: PresetMiniOptions) => presetMini(options),
-  })
-})()
+window.__unocss_presets = Object.assign(window.__unocss_presets ?? {}, {
+  presetMini: (options: PresetMiniOptions) => presetMini(options),
+})
