+++ packages/runtime/src/presets/preset-web-fonts.ts
@@ -1,94 +1,8 @@
-import type { Preset } from '@unocss/core'
-import { toArray } from '@unocss/core'
 import type { WebFontsOptions } from '@unocss/preset-web-fonts'
-import { normalizedFontMeta, providers } from '@unocss/preset-web-fonts'
+import { createPreset } from '../../../preset-web-fonts/src/preset'
 
-(() => {
-  const $fetch = (url: string) => fetch(url).then(data => data.text())
+const presetWebFonts = createPreset(url => fetch(url).then(data => data.json()))
 
-  const presetWebFonts = (options: WebFontsOptions = {}): Preset<any> => {
-    const {
-      provider: defaultProvider = 'google',
-      extendTheme = true,
-      inlineImports = true,
-      themeKey = 'fontFamily',
-      customFetch,
-    } = options
-
-    const fontObject = Object.fromEntries(
-      Object.entries(options.fonts || {})
-        .map(([name, meta]) => [name, toArray(meta).map(m => normalizedFontMeta(m, defaultProvider))]),
-    )
-    const fonts = Object.values(fontObject).flatMap(i => i)
-
-    const importCache: Record<string, Promise<string>> = {}
-
-    async function importUrl(url: string) {
-      if (inlineImports) {
-        if (!importCache[url]) {
-          const promise = (customFetch || $fetch)(url)
-          importCache[url] = promise.catch((e) => {
-            console.error('Failed to fetch web fonts')
-            console.error(e)
-            if (typeof process !== 'undefined' && process.env.CI)
-              throw e
-          })
-        }
-        return await importCache[url]
-      }
-      else {
-        return `@import url('${url}')`
-      }
-    }
-
-    const preset: Preset<any> = {
-      name: '@unocss/preset-web-fonts',
-      preflights: [
-        {
-          async getCSS() {
-            const names = new Set(fonts.map(i => i.provider || defaultProvider))
-            const preflights: (string | undefined)[] = []
-
-            for (const name of names) {
-              const fontsForProvider = fonts.filter(i => i.provider === name)
-              const provider = providers[name]
-
-              if (provider.getImportUrl) {
-                const url = provider.getImportUrl(fontsForProvider)
-                if (url)
-                  preflights.push(await importUrl(url))
-              }
-
-              preflights.push(provider.getPreflight?.(fontsForProvider))
-            }
-
-            return preflights.filter(Boolean).join('\n')
-          },
-        },
-      ],
-    }
-
-    if (extendTheme) {
-      preset.extendTheme = (theme) => {
-        if (!theme[themeKey])
-          theme[themeKey] = {}
-        const obj = Object.fromEntries(
-          Object.entries(fontObject)
-            .map(([name, fonts]) => [name, fonts.map(f => providers[f.provider || defaultProvider].getFontName(f))]),
-        )
-        for (const key of Object.keys(obj)) {
-          if (typeof theme[themeKey][key] === 'string')
-            theme[themeKey][key] = obj[key].map(i => `${i},`).join('') + theme[themeKey][key]
-          else
-            theme[themeKey][key] = obj[key].join(',')
-        }
-      }
-    }
-
-    return preset
-  }
-
-  window.__unocss_presets = Object.assign(window.__unocss_presets ?? {}, {
-    presetWebFonts: (options: WebFontsOptions) => presetWebFonts(options),
-  })
-})()
+window.__unocss_presets = Object.assign(window.__unocss_presets ?? {}, {
+  presetWebFonts: (options: WebFontsOptions) => presetWebFonts(options),
+})
